pyproject.toml
[tool.poetry]
name = "guobang_bot"
version = "0.1.0"
description = "guobang_bot"
authors = ["nb adapter install nonebot-adapter-onebot"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["guobang_bot/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
